respond to PR feedback
Pull static file serving and redirection out of FileSystemAPI.
